+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 - Updated Umpire::Allocator 'deallocate' documentation brief on handling
   deallocations on nullptr.
 
+- Benchmark that overlooks overhead of malloc calls with No-Op memory resource.
+
 ### Changed
 
 - Organized configuration options in config.hpp.in in alphabetical order